requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
feedparser>=6.0.10
requests>=2.31.0
python-dateutil>=2.8.2
matplotlib>=3.8.2
wordcloud>=1.9.3
plotly>=5.18.0
vaderSentiment>=3.3.2
transformers>=4.39.0
torch>=2.2.0
python-dotenv>=1.0.0